Clarifications
Hi!

I'm currently making my way through(haha) the Stream tutorial, and am proposing these changes in text because, as a total n00b, I found the original wording to be very confusing.  I rearranged the order on some of the instructions, and broke a couple of hte concepts down more explicitly to be more clear for beginners.  Thanks so much!