Fix verification of HTTP SERVER exercise
Hey there,

The `setup.js` incorrectly binds both the submission and solution clients on a single port (`aPort` instead of `port`), which *should* screw up for everyone, everytime, on `verify`.  This seems to work, unexpectedly, for most people (can't fathom why), but it does break for several people (which is should!).

Just checking on the proper port helps.