Update Secretz solution for latest tar. Fixes #202.
The current solution for the 'Secretz' adventure fails with the error described in #202. Using the `new` keyword will fix this particular error, but the solution will still fail because of API changes to the `tar` parser events. Per the [documentation](https://github.com/npm/node-tar/blob/master/README.md#class-tarparse), we must now call `e.resume()` so the stream will continue to flow.